feat: separate test into small file
